requirements.txt
fastapi
uvicorn
motor==3.3.2
pymongo[srv]==4.6.1
python-dotenv
google-genai
bcrypt
pyjwt